+++ CHANGELOG.md
@@ -6,9 +6,11 @@
     - Error Handling with Try...Catch
     - Regular Expressions [Part 1] - Evaluation Functions
     - Regular Expressions [Part 2] - Metacharacter Symbols
+    - Regular Expressions [Part 3] - Character Sets & Quantifiers
 - An example with error catching was added.
 - An example with evaluation functions was added.
 - An example with metacharacter symbols was added.
+- An example with character sets and quantifiers was added.
 
 ## September 21, 2020
 - Modern JavaScript From The Beginning (Udemy)